Add an extra check for invalid exponents
`fast_float` allows for exponents containing floats, which are not
allowed per the JSON5 spec.

Constraining `fast_float` to a format which does not allow for these
exponents results in other supported usages being rejected.

To handle this, explicitly check by adding a new guard.
`has_invalid_exponent` is a string scan which returns true if it finds
a suffix on the string which looks like an invalid exponent.